backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.1
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0

psycopg2==2.9.3
redis==4.3.1
reportlab==3.6.9
requests==2.27.1
python-decouple==3.5